+++ programs/alpha-vault/src/instructions/create_escrow/mod.rs
@@ -1,7 +1,0 @@
-pub mod create_new_escrow;
-pub mod create_permissioned_escrow;
-pub mod create_permissioned_escrow_with_authority;
-
-pub use create_new_escrow::*;
-pub use create_permissioned_escrow::*;
-pub use create_permissioned_escrow_with_authority::*;