+++ programs/alpha-vault/src/instructions/fill_amm/mod.rs
@@ -1,5 +1,0 @@
-pub mod fill_dlmm;
-pub use fill_dlmm::*;
-
-pub mod fill_dynamic_amm;
-pub use fill_dynamic_amm::*;