+++ programs/alpha-vault/src/instructions/handle_crank_fee_whitelist/mod.rs
@@ -1,5 +1,0 @@
-pub mod close_crank_fee_whitelist;
-pub mod create_crank_fee_whitelist;
-
-pub use close_crank_fee_whitelist::*;
-pub use create_crank_fee_whitelist::*;