+++ ts-client/src/examples/createCustomizableDlmmWithPermissionlessVault.ts
@@ -55,6 +55,7 @@ async function createCustomizableDlmmWithPermissionlessVault(
     hasAlphaVault,
     creator,
     activationPoint,
+    false,
     {
       cluster: "devnet",
     }